jpa updates
